+++ src/hydro/diffusion/diffusion.cpp
@@ -1,6 +1,6 @@
 //========================================================================================
 // AthenaPK - a performance portable block structured AMR astrophysical MHD code.
-// Copyright (c) 2021, Athena-Parthenon Collaboration. All rights reserved.
+// Copyright (c) 2021-2023, Athena-Parthenon Collaboration. All rights reserved.
 // Licensed under the 3-clause BSD License, see LICENSE file for details
 //========================================================================================
 //! \file diffusion.cpp
@@ -25,6 +25,28 @@ TaskStatus CalcDiffFluxes(StateDescriptor *hydro_pkg, MeshData<Real> *md) {
       ThermalFluxIsoFixed(md);
     } else {
       ThermalFluxGeneral(md);
+    }
+  }
+  const auto &viscosity = hydro_pkg->Param<Viscosity>("viscosity");
+  if (viscosity != Viscosity::none) {
+    const auto &mom_diff = hydro_pkg->Param<MomentumDiffusivity>("mom_diff");
+
+    if (viscosity == Viscosity::isotropic &&
+        mom_diff.GetCoeffType() == ViscosityCoeff::fixed) {
+      MomentumDiffFluxIsoFixed(md);
+    } else {
+      MomentumDiffFluxGeneral(md);
+    }
+  }
+  const auto &resistivity = hydro_pkg->Param<Resistivity>("resistivity");
+  if (resistivity != Resistivity::none) {
+    const auto &ohm_diff = hydro_pkg->Param<OhmicDiffusivity>("ohm_diff");
+
+    if (resistivity == Resistivity::ohmic &&
+        ohm_diff.GetCoeffType() == ResistivityCoeff::fixed) {
+      OhmicDiffFluxIsoFixed(md);
+    } else {
+      OhmicDiffFluxGeneral(md);
     }
   }
   return TaskStatus::complete;